reafactor: move PostCommentWithId fix: remove filter from postdetail kpis
