+++ browser-extension/src/entrypoints/posts/Report/ReportGroupingUtils.tsx
@@ -1,4 +1,3 @@
-import { PostCommentWithId } from "../Posts/CommentsTable";
 import { Post } from "@/shared/model/post/Post";
 import React from "react";
 import { ReportOrganizationType } from "./Stepper/BuildReport";
@@ -7,7 +6,7 @@ import {
   getTitlePublicationHeader,
   LABEL_URL,
 } from "./reportData";
-import { buildPostKey } from "@/shared/utils/post-util";
+import { buildPostKey, PostCommentWithId } from "@/shared/utils/post-util";
 
 export interface GroupedData {
   groupKey: string;
